+++ R/reexport.R
@@ -1,0 +1,7 @@
+##' @importFrom ggplotify as.grob
+##' @export
+ggplotify::as.grob
+
+##' @importFrom ggplotify as.ggplot
+##' @export
+ggplotify::as.ggplot